After refactoring
Use of memory optimized, time optimized

+++ common.h
@@ -1,8 +1,8 @@
 #pragma once
 
-const int max_len_of_filename = 50;
+const size_t max_len_of_filename = 50;
 
 struct String {
-    char* ptr;
-    int   len;
+    char*   ptr;
+    size_t  len;
 };
